# perf(circom): optimize zkey wtns reading