… vendors/circom/benchmark/rapidsnark_runner.h
… vendors/circom/benchmark/runner.h
… vendors/circom/benchmark/tachyon_runner.h
… vendors/circom/circomlib/zkey/BUILD.bazel
… vendors/circom/circomlib/zkey/zkey.h
… vendors/circom/prover_main.cc