# Description

Zkey and wtns files are now read using `mmap()` and is much faster!

Before

```
Start parsing zkey
Time taken for parsing zkey: 5.97333 s
Start parsing witness
Time taken for parsing witness: 0.319247 s
```

After
```
Start parsing zkey
Time taken for parsing zkey: 0.611177 s
Start parsing witness
Time taken for parsing witness: 0.097334 s
```