@chokobole 2024-07-29
> LGTM please fix ci lint~

CI lint asks me to fix the code from chromium, so I'll leave it as it is for easy maintenance.